requirements.txt
django==4.0
django-ckeditor==6.2.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
django-filter==21.1
mysqlclient==2.1.0
markdown==3.3.6